send the object upwards
So here, I have an onclick and I also know that since this component is being rendered from a loop, whatever card I click will be unique. Since it is unique I can pass it in a function and use the data in the modal. That way whatever modal is clicked will have the right data.