lots of red
So most of the red you see here is not what I have deleted, it is because I had a code formatted plugin.

Remember that dispatch sends an >>> action <<<, and an action is a function.

Here you are trying to send an action as well as another function called setCurrentTask.